bump parallel_hashmap to v2.0.0

+++ src/parallel_hashmap/phmap_bits.h
@@ -18,7 +18,7 @@
 //
 // Includes work from abseil-cpp (https://github.com/abseil/abseil-cpp)
 // with modifications.
-// 
+//
 // Copyright 2018 The Abseil Authors.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -51,7 +51,7 @@
 #include "phmap_config.h"
 
 #ifdef _MSC_VER
-    #pragma warning(push)  
+    #pragma warning(push)
     #pragma warning(disable : 4514) // unreferenced inline function has been removed
 #endif
 
@@ -183,7 +183,7 @@ inline void UnalignedStore64(void *p, uint64_t v) { memcpy(p, &v, sizeof v); }
 
 #ifdef PHMAP_HAVE_INTRINSIC_INT128
     __extension__ typedef unsigned __int128 phmap_uint128;
-    inline uint64_t umul128(uint64_t a, uint64_t b, uint64_t* high) 
+    inline uint64_t umul128(uint64_t a, uint64_t b, uint64_t* high)
     {
         auto result = static_cast<phmap_uint128>(a) * static_cast<phmap_uint128>(b);
         *high = static_cast<uint64_t>(result >> 64);
@@ -193,7 +193,7 @@ inline void UnalignedStore64(void *p, uint64_t v) { memcpy(p, &v, sizeof v); }
 #elif (defined(_MSC_VER))
     #if defined(_M_X64)
         #pragma intrinsic(_umul128)
-        inline uint64_t umul128(uint64_t a, uint64_t b, uint64_t* high) 
+        inline uint64_t umul128(uint64_t a, uint64_t b, uint64_t* high)
         {
             return _umul128(a, b, high);
         }
@@ -276,7 +276,7 @@ PHMAP_BASE_INTERNAL_FORCEINLINE uint32_t CountLeadingZeros64Slow(uint64_t n) {
     if (n >> 16) zeroes -= 16, n >>= 16;
     if (n >> 8) zeroes -= 8, n >>= 8;
     if (n >> 4) zeroes -= 4, n >>= 4;
-    return "\4\3\2\2\1\1\1\1\0\0\0\0\0\0\0"[n] + zeroes;
+    return (uint32_t)("\4\3\2\2\1\1\1\1\0\0\0\0\0\0\0"[n] + zeroes);
 }
 
 PHMAP_BASE_INTERNAL_FORCEINLINE uint32_t CountLeadingZeros64(uint64_t n) {
@@ -284,17 +284,17 @@ PHMAP_BASE_INTERNAL_FORCEINLINE uint32_t CountLeadingZeros64(uint64_t n) {
     // MSVC does not have __buitin_clzll. Use _BitScanReverse64.
     unsigned long result = 0;  // NOLINT(runtime/int)
     if (_BitScanReverse64(&result, n)) {
-        return (int)(63 - result);
+        return (uint32_t)(63 - result);
     }
     return 64;
 #elif defined(_MSC_VER) && !defined(__clang__)
     // MSVC does not have __buitin_clzll. Compose two calls to _BitScanReverse
     unsigned long result = 0;  // NOLINT(runtime/int)
     if ((n >> 32) && _BitScanReverse(&result, (unsigned long)(n >> 32))) {
-        return 31 - result;
+        return  (uint32_t)(31 - result);
     }
     if (_BitScanReverse(&result, (unsigned long)n)) {
-        return 63 - result;
+        return (uint32_t)(63 - result);
     }
     return 64;
 #elif defined(__GNUC__) || defined(__clang__)
@@ -309,7 +309,7 @@ PHMAP_BASE_INTERNAL_FORCEINLINE uint32_t CountLeadingZeros64(uint64_t n) {
     if (n == 0) {
         return 64;
     }
-    return __builtin_clzll(n);
+    return  (uint32_t)__builtin_clzll(n);
 #else
     return CountLeadingZeros64Slow(n);
 #endif
@@ -658,7 +658,7 @@ inline void Store64(void *p, uint64_t v) {
 }  // namespace phmap
 
 #ifdef _MSC_VER
-     #pragma warning(pop)  
+     #pragma warning(pop)
 #endif
 
 #endif // phmap_bits_h_guard_
